requirements.txt
pywin32==306
Pillow==10.4.0
python-barcode==0.15.1
